requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
coverage==7.11.0
    # via -r requirements.in
distro==1.9.0
    # via openai
fastapi==0.116.1
    # via -r requirements.in
filetype==1.2.0
    # via langchain-google-genai
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.26.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.185.0
    # via google-generativeai
google-auth==2.41.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via langchain-google-genai
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.4
    # via sqlalchemy
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.11.1
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via -r requirements.in
langchain-core==0.3.79
    # via
    #   langchain
    #   langchain-google-genai
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==2.0.10
    # via -r requirements.in
langchain-openai==0.3.29
    # via -r requirements.in
langchain-text-splitters==0.3.11
    # via langchain
langgraph==0.6.4
    # via -r requirements.in
langgraph-checkpoint==2.1.2
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.6.5
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.37
    # via
    #   langchain
    #   langchain-core
markdown==3.9
    # via -r requirements.in
numpy==2.3.4
    # via pandas
openai==1.109.1
    # via langchain-openai
orjson==3.11.3
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
pandas==2.3.3
    # via -r requirements.in
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   google-generativeai
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.5
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   langchain
    #   langchain-core
regex==2025.10.23
    # via tiktoken
requests==2.32.5
    # via
    #   google-api-core
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.44
    # via langchain
starlette==0.47.3
    # via fastapi
tenacity==9.1.2
    # via langchain-core
tiktoken==0.12.0
    # via langchain-openai
tqdm==4.67.1
    # via
    #   google-generativeai
    #   openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   google-generativeai
    #   grpcio
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
